# web/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ASG Card API",
    "version": "0.5.0",
    "description": "ASG Card virtual card issuance API for AI agents. Supports two payment rails: x402 with USDC on Stellar mainnet (agent-autonomous) and Stripe Machine Payments Protocol (owner-in-the-loop). See https://asgcard.dev/docs for integration guide."
  },
  "servers": [
    {
      "url": "https://api.asgcard.dev",
      "description": "ASG Card API — Stellar Mainnet + Stripe MPP"
    }
  ],
...
  "paths": {
...
    "/cards/create/tier/{amount}": {
      "post": {
        "tags": [
          "x402"
        ],
        "summary": "Create a virtual card ($5–$5,000)",
        "description": "x402 payment flow. Clients receive HTTP 402 with a payment challenge (stellar:pubnet, exact scheme), then retry with X-PAYMENT header containing signed Stellar transaction. The 201 response includes a `detailsEnvelope` with full PAN, CVV, expiry, billing address, and one-time-access metadata (agent-first model). PAN/CVV are never logged or exposed in metrics.",
        "operationId": "createCardByTier",
        "parameters": [
          {
            "name": "amount",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Amount in USD ($5–$5,000)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCardRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Card created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCardResponse"
                }
              }
            }
          },
          "402": {
            "description": "x402 payment challenge (challenge schema may vary by facilitator)"
          }
        }
      }
    },
    "/cards/fund/tier/{amount}": {
      "post": {
        "tags": [
          "x402"
        ],
        "summary": "Fund an existing card ($5–$5,000)",
        "description": "This endpoint participates in the x402 payment flow.",
        "operationId": "fundCardByTier",
        "parameters": [
          {
            "name": "amount",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Amount in USD ($5–$5,000)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundCardRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Card funded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundCardResponse"
                }
              }
            }
          },
          "402": {
            "description": "x402 payment challenge"
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "CreateCardRequest": {
        "type": "object",
        "required": [
          "nameOnCard",
          "email"
        ],
        "properties": {
          "nameOnCard": {
            "type": "string",
            "minLength": 1,
            "example": "AI AGENT"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "agent@example.com"
          }
        }
      },
      "PaymentInfo": {
        "type": "object",
        "properties": {
          "amountCharged": {
            "type": "number",
            "format": "float",
            "example": 20.35
          },
          "txHash": {
            "type": "string"
          },
          "network": {
            "type": "string",
            "example": "stellar"
          }
        }
      },
      "BillingAddress": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "zip": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        }
      },
...
      "CardSummary": {
        "type": "object",
        "properties": {
          "cardId": {
            "type": "string",
            "format": "uuid"
          },
          "nameOnCard": {
            "type": "string"
          },
          "balance": {
            "type": "number",
            "format": "float"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateCardResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "card": {
            "$ref": "#/components/schemas/CardSummary"
          },
          "payment": {
            "$ref": "#/components/schemas/PaymentInfo"
          },
          "detailsEnvelope": {
            "type": "object",
            "description": "Sensitive card details with one-time access metadata (agent-first model)",
            "properties": {
              "cardNumber": {
                "type": "string"
              },
              "expiryMonth": {
                "type": "integer"
              },
              "expiryYear": {
                "type": "integer"
              },
              "cvv": {
                "type": "string"
              },
              "billingAddress": {
                "$ref": "#/components/schemas/BillingAddress"
              },
              "oneTimeAccess": {
                "type": "boolean",
                "description": "Indicates this envelope is only returned on creation"
              },
              "expiresInSeconds": {
                "type": "integer",
                "description": "Seconds before the caller should consider these details expired"
              },
              "note": {
                "type": "string"
              }
            }
          }
        }
      },
      "FundCardRequest": {
        "type": "object",
        "required": [
          "cardId"
        ],
        "properties": {
          "cardId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "FundCardResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "cardId": {
            "type": "string",
            "format": "uuid"
          },
          "fundedAmount": {
            "type": "number",
            "format": "float"
          },
          "newBalance": {
            "type": "number",
            "format": "float"
          },
          "payment": {
            "$ref": "#/components/schemas/PaymentInfo"
          }
        }
      },
...
    }
  },
...
}